refactor: sifli_debug也是SF32LB52特有实现
